Consume API endpoint for updating an event
1. Implement functionality for switching center image field between required and not-required for creating or updating center respectively
2. Implement functionality to switch title of addEvent and addCenter modals between create and modify states
3.  Implement functionality to bind eventDetails and centerDetails state props to event and center form fields upon update action
4. create page actions for setting modal title, required field property, eventDefaults and centerDefaults and implement reducer handler
5. In manageDetailsHeader component, implement feature to dispatch setModalTitle, setRequired, setEventDefaults, and setCenterDefaults actions for either event or center update contexts
6. Prompt for validation of delete action in when user tries to delete an event/center
7.  In navBar component, implement feature to dispatch setModalTitle, setRequired, setEventDefaults, and setCenterDefaults actions for new event or center creation contexts
8. In event and center server Controllers: validate for picture field (center) and delete/relace picture file when an event/center image was updated in put http action
9. Remove validation for center picture in validator middleware

+++ Client/src/components/manageDetailsHeader.js
@@ -3,6 +3,7 @@ import editIcon from '../resources/images/glyphicons-151-edit.png';
 import removeIcon from '../resources/images/glyphicons-17-bin.png';
 import { connect } from 'react-redux';
 import axios from 'axios';
+import { setModalTitle, setRequired, setEventDefaults, setCenterDefaults } from '../actions/pageActions'
 
 const mapStateToProps = state => {
   let currentPage = state.page.currentPage;
@@ -12,14 +13,22 @@ const mapStateToProps = state => {
     owner: owner,
     currentPage: state.page.currentPage,
     eventDetails: state.events.event,
-    centerDetails: state.centers.centerDetails    
+    centerDetails: state.centers.centerDetails
+  };
+};
+const mapDispatchToProps = dispatch => {
+  return {
+    setModalTitle: title => dispatch(setModalTitle(title)),
+    setRequired: value => dispatch(setRequired(value)),
+    setEventDefaults: data => dispatch(setEventDefaults(data)),
+    setCenterDefaults: data => dispatch(setCenterDefaults(data))
   };
 };
 const Manager = (props) => {
   const content = (
     <td>
       <div className="manage">
-        <button type="submit" id="editEvent" className="btn btn-success" data-toggle="modal" data-target={props.editModal}>
+        <button onClick={props.setModalProps} type="submit" id="editEvent" className="btn btn-success" data-toggle="modal" data-target={props.editModal}>
           <img src={editIcon} alt="Edit" />
         </button>
         <button onClick={props.deleteEvent} type="submit" className="btn btn-danger icon-margin-left" id="deleteEvent">
@@ -37,24 +46,34 @@ class ManageDetails extends Component {
     this.delete = this.delete;
   }
   delete() {
-    const urlGenerator = () => {
-      const type = (currentPage === 'manageEvent') ? 'events' : 'centers';
-      return `http://localhost:8080/api/v1/${type}/${param}/?token=${JSON.parse(localStorage.token).value}&file=${resource.picture}`;
+    const validate = confirm('Confirm delete action?');
+    if (validate) {
+      const urlGenerator = () => {
+        const type = (currentPage === 'manageEvent') ? 'events' : 'centers';
+        return `http://localhost:8080/api/v1/${type}/${param}/?token=${JSON.parse(localStorage.token).value}&file=${resource.picture}`;
+      }
+      const url = urlGenerator();
+      axios.delete(url).then(res => {
+        alert(res.data.data)
+        history.push('/dashboard/events');
+      }).catch(err => {
+        alert(err.response.data.message);
+        (err.response.status === 403 || err.response.status === 401) && logout('addNewEvent', history);
+      });
     }
-    const url = urlGenerator();
-    axios.delete(url).then(res => {
-      alert(res.data.data)
-      history.push('/dashboard/events');
-    }).catch(err => {
-      alert(err.response.data.message);
-      (err.response.status === 403 || err.response.status === 401) && logout('addNewEvent', history);
-    });
   }
   render() {
     param = this.props.param;
     currentPage = this.props.currentPage;
     history = this.props.history;
     resource = (currentPage === 'manageEvent') ? this.props.eventDetails[0] : this.props.centerDetails[0];
+
+    const setModalProps = title => {
+      this.props.setModalTitle(title);
+      this.props.setRequired(false);
+      (title === 'Modify Event') ? this.props.setEventDefaults(this.props.eventDetails[0])
+        : this.props.setCenterDefaults(this.props.centerDetails[0]);
+    }
 
     const content = (
       <table className="table-responsive col-sm-12 bg-transparent  zero-padding">
@@ -63,7 +82,8 @@ class ManageDetails extends Component {
             <td style={{ width: '80%' }}><h4 className="text-center"><b>{this.props.title}</b></h4></td>
             {
               (JSON.parse(localStorage.token).id === this.props.owner)
-              && <Manager deleteEvent={this.delete} editModal={this.props.editModal} />
+              && <Manager setModalProps={() => (currentPage === 'manageEvent') ? setModalProps('Modify Event')
+                : setModalProps('Modify Center')} deleteEvent={this.delete} editModal={this.props.editModal} />
             }
           </tr>
         </tbody>
@@ -72,4 +92,4 @@ class ManageDetails extends Component {
     return content;
   }
 }
-export const ManageDetailsHeader = connect(mapStateToProps)(ManageDetails);
+export const ManageDetailsHeader = connect(mapStateToProps, mapDispatchToProps)(ManageDetails);
